Rename: cas -> atomic
