+++ concurrency-practice/src/main/java/me/staek/atomic/_03_AtomicBoolean.java
@@ -1,4 +1,4 @@
-package me.staek.cas;
+package me.staek.atomic;
 
 import java.util.concurrent.atomic.AtomicBoolean;
