+++ concurrency-practice/src/main/java/me/staek/atomic/_06_AtomicReference_API.java
@@ -1,4 +1,4 @@
-package me.staek.cas;
+package me.staek.atomic;
 
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.UnaryOperator;
