+++ concurrency-practice/src/main/java/me/staek/atomic/_07_AtomicReference_API2.java
@@ -1,4 +1,4 @@
-package me.staek.cas;
+package me.staek.atomic;
 
 import java.util.concurrent.atomic.AtomicReference;
